requirements.txt
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
loguru==0.6.0